% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibo.R
\name{sensibo}
\alias{sensibo}
\alias{sensibo_call}
\alias{sensibo_info}
\alias{sensibo_devices}
\alias{sensibo_maintain_fan_speed}
\alias{sensibo_set_fan_speed}
\alias{sensibo_config}
\alias{sensibo_maintain_devices}
\alias{sensibo_historical_measurements}
\title{Sensibo}
\usage{
sensibo_call(
  path,
  ...,
  query = list(),
  verb = "GET",
  api_key = Sys.getenv("SENSIBO_API_KEY")
)

sensibo_info()

sensibo_devices()

sensibo_maintain_fan_speed(
  device_id,
  cur_temp,
  min_temp,
  max_temp,
  curr_fan_speed,
  cool_fan_speed = "medium_low",
  norm_fan_speed = "medium",
  warm_fan_speed = "high"
)

sensibo_set_fan_speed(device_id, fan_speed)

sensibo_config()

sensibo_maintain_devices()

sensibo_historical_measurements(device_id, days = 1L)
}
\arguments{
\item{device_id}{device ID from sensibo}

\item{cur_temp}{current temperature reading}

\item{min_temp}{minimum temperature allowed}

\item{max_temp}{maximum temperature allowed}

\item{curr_fan_speed}{current fan speed}

\item{cool_fan_speed}{fan speed if the room is too cool}

\item{norm_fan_speed}{fan speed if the room is within range}

\item{warm_fan_speed}{fan speed if the room is too warm}

\item{fan_speed}{fan speed string setting}

\item{days}{number of days back to retrieve}
}
\description{
Sensibo is a device that allows control of air conditioning units.

API documentation is \href{https://sensibo.github.io/}{here}.
}
\section{Functions}{
\itemize{
\item \code{sensibo_call}: Generic API call

\item \code{sensibo_info}: Get all the data as a tibble

\item \code{sensibo_devices}: Parse sensibo information into usable format

\item \code{sensibo_maintain_fan_speed}: maintains the temperature by adjusting the fan speed

\item \code{sensibo_set_fan_speed}: Sets the fan speed of a sensibo device

\item \code{sensibo_config}: configuration for the maintenance job

\item \code{sensibo_maintain_devices}: maintains devices according to the configuration

\item \code{sensibo_historical_measurements}: fetch historical data
}}

