% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjhome-package.R
\docType{package}
\name{tjhome-package}
\alias{tjhome}
\alias{tjhome-package}
\title{tjhome: Home Automation for TJ Palanca's Home}
\description{
I use this package to manage the smart devices in my home.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.tjpalanca.com/tjhome}
  \item \url{https://github.com/tjpalanca/tjhome}
  \item Report bugs at \url{https://github.com/tjpalanca/tjhome/issues}
}

}
\author{
\strong{Maintainer}: Troy James Palanca \email{mail@tjpalanca.com} (\href{https://orcid.org/0000-0002-0610-9213}{ORCID})

}
\keyword{internal}
