% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_maintain_ac_temp}
\alias{job_maintain_ac_temp}
\title{Job: Maintain AC Temperature}
\usage{
job_maintain_ac_temp(interval = 5)
}
\arguments{
\item{interval}{polling interval in seconds}
}
\description{
This keeps the temperature maintained across a particular range.
}
